app.py/resume-to-job-matcher/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.35.27
botocore==1.35.27
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
cohere==5.10.0
colorama==0.4.6
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docopt==0.6.2
einops==0.8.0
executing==2.1.0
fastavro==1.9.7
fastjsonschema==2.20.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
google-ai-generativelanguage==0.6.6
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.0
idna==3.10
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.3.0
langchain-cohere==0.3.0
langchain-community==0.3.0
langchain-core==0.3.5
langchain-experimental==0.3.1.post1
langchain-google-genai==2.0.0
langchain-mongodb==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.125
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==4.25.5
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.9.1
pyparsing==3.1.4
pypdf==5.0.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yarl==1.11.1